vector-verify: add generated vec_pack_i16 + int-cast error prints
Third Half-B capture kernel: native VPACK, int16 -> int8. Empirically resolved
the open saturation question -- native v.pack<int8_t>() TRUNCATES (low 8 bits)
regardless of the saturation mode, so the kernel matches the Half-A
golden slice with sat=false (128->-128, 256->0, ...). EMU-smoke PASS (chess),
no new interpreter gap -- the emulator already models truncating pack correctly.
This is genuine distinct coverage: srs verifies the saturating acc->vector
narrow, pack verifies the truncating vector->vector native narrow.

Generator fix: the host-harness error print now casts IN/got/EXP to int, so
int8 (char-typed) outputs print as readable integers instead of raw bytes (the
int8 mismatch that diagnosed the truncation was unreadable without this).
Regenerated vec_ups_i32 to pick up the same print fix (idempotent; still PASS).

Generated using Claude Code.

+++ tools/gen_vector_kernel.py
@@ -186,8 +186,8 @@ int main(int argc, const char *argv[]) {
   for (int i = 0; i < N; i++) {
     if (bufOut[i] != EXP[i]) {
       if (errors < 10)
-        std::cout << "Error [" << i << "]: in=" << IN[i]
-                  << " got=" << bufOut[i] << " != exp=" << EXP[i] << "\\n";
+        std::cout << "Error [" << i << "]: in=" << (int)IN[i]
+                  << " got=" << (int)bufOut[i] << " != exp=" << (int)EXP[i] << "\\n";
       errors++;
     }
   }
